+++ src/modules/ZenSelection/ZenSelection.tsx
@@ -10,7 +10,7 @@ import { getUnlockableLevelTypes } from "@/game/level-types";
 import { LevelType } from "@/game/level-types/types";
 import { timesMap } from "@/support/timeMap";
 
-import { DIFFICULTY_LEVELS } from "../GameModi/zenModeConstants";
+import { DIFFICULTY_LEVELS } from "../GameModi/constants";
 import { BackgroundContext } from "../Layout/BackgroundContext";
 import { BetaContext } from "../Layout/BetaContext";
 
@@ -54,7 +54,7 @@ export const ZenSelection: React.FC<Props> = ({
 
   return (
     <div className="flex h-full flex-col items-center">
-      <div className="pt-safeTop flex w-full flex-row gap-x-2 pb-2 pl-safeLeft pr-safeRight">
+      <div className="flex w-full flex-row gap-x-2 pb-2 pl-safeLeft pr-safeRight pt-safeTop">
         {onOpenSettings && (
           <TopButton buttonType="settings" onClick={onOpenSettings} />
         )}