+++ src/ui/TransparentButton/TransparentButton.tsx
@@ -16,8 +16,8 @@ export const TransparentButton: React.FC<Props> = ({
   size = "default"
 }) => {
   const classNames = clsx(
-    "inline-block rounded-full border border-black p-2 shadow-md",
-    "text-center align-middle text-black",
+    "inline-block select-none rounded-full border border-black p-2 shadow-md",
+    "text-center align-middle text-black focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-black",
     "transition-transform active:scale-90",
     {
       "px-4 text-sm": size === "default"