+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 1.26.0 - 2024-11-24
+
+### Added
+
+- Auto update theme on activation date
+- More preparations for spring theme
+
+### Changed
+
+- Improved indication of difficulty increase on level track
+- Updated Share mechanic, added share through QR
+
+### Fixed
+
+- Fixed stuttering on start of level animation
+
 ## Version 1.25.1 - 2024-11-19
 
 ### Fixed